pyproject.toml
[tool.pdm]

[tool.pdm.dev-dependencies]
dev = [
    "ruff==0.0.264",
    "pytest~=7.3.1",
    "pytest-asyncio~=0.21.0",
    "autopep8~=2.0.2",
]

[tool.pdm.build]
package-dir = "src"
#excludes = [".gdf", ".venv"]


[project]
name = "nats-nsc"
version = "0.2.2"
description = "Python wrapper for nsc, nats credential manager"
authors = [
    {name = "Mikołaj Nowak", email = "12396461+m3nowak@users.noreply.github.com"},
]
dependencies = [
    "pyJWT~=2.6.0",
    "aiofiles~=23.1.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.autopep8]
max_line_length = 120

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
pythonpath = [
  "src/"
]
